HACK: restore: Bypass failure during SELinux context parsing
On rare occasions (about 1 : 20), `ls` may return a non-zero code,
causing bash to terminate the shell. The issue appears to be
caused by a second party operating on the same working directory
at the sime time that we execute the `ls` command.

Add `|| true` to the end of the pipeline to bypass the shell
termination. This should be safe to implement since there is no
scenario in which `$data_dir` is empty, nor is there a possibility
that `ls`, `head`, or `cut` can return an error code.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>